requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiopg==1.4.0
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dictdiffer==0.9.0
distro==1.8.0
Flask==3.0.0
Flask-Admin==1.6.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
magic-filter==1.0.12
MarkupSafe==2.1.3
multidict==6.0.4
openai==1.6.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
tomlkit==0.12.3
tortoise-orm==0.20.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
watchdog==3.0.0
Werkzeug==3.0.1
WTForms==3.1.1
yarl==1.9.4
